requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
docker-pycreds==0.4.0
dotenv==0.9.9
eval_type_backport==0.2.2
exceptiongroup==1.2.1
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.5.2
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
openai==1.34.0
packaging==24.1
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.8
protobuf==6.30.2
psutil==7.0.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.27.0
setproctitle==1.3.6
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.6.0
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
zipp==3.21.0
